Refactor to support native aws-sdk
This is a fairly large change to the way files are handled in this.

We have continued to have issues with jclouds and want to move away from
it for the core cloud providers (but may continue to use it for
backwards compatibility)

To do this, we introduce a new small interface, BlobWriter, and
implement two classes, JCloudsBlobWriter and S3BlobsWriter, the
JCloudsBlobWriter should be entirely backwards compatbile, and we
instead introduce a new "provider", "s3v2" that uses the native AWS SDK.

This is signifcantly simpler than jclouds and I suggest we do the same
for GCS and Azure.

This currently should not be any breaking changes, though it does
slightly relax the config validation around S3 fields