fix: cleanup the current batch after flush crashed (#824)
In the existing logic, whether the flush operation fails or succeeds, the variables `currentBatchSize` and `currentBatchBytes` will be cleared. However, if an error such as an Out of Memory (OOM) exception occurs during the flush operation, it will be caught here:
https://github.com/streamnative/pulsar-io-cloud-storage/blob/06d8c33afa7d813ffc2019ed4331a3883f9f9044/src/main/java/org/apache/pulsar/io/jcloud/sink/BlobStoreAbstractSink.java#L217-L221

Nonetheless, the batch might not be cleared properly. This could result in an empty queue, despite `currentBatchSize` and `currentBatchBytes` not being reset to 0.

- The batch is now cleared before the actual flush operation takes place.

(cherry picked from commit f2a2a2f27028cc27308bdcb0cd7f3e548e6f5512)