Add default value for the partitionType to `partition` (#863)
### Motivation

In https://github.com/streamnative/pulsar-io-cloud-storage/pull/845, we have deprecated `partitionerType` and recommand users to use `partitioner`.

Set the default value of `partitionerType` to `partition`. So that we could be able to not configure it. Otherwise, it will raise exceptions if we don't configure it.

### Modifications

- Add default value for the partitionType to `partition`

(cherry picked from commit ba24f85c922a61cef14466a913ebd408424dd7cf)